On crée les méthodes des services

+++ Craft2Gether/src/main/java/fr/crafttogether/services/BlocService.java
@@ -1,5 +1,6 @@
 package fr.crafttogether.services;
 
+import fr.crafttogether.exceptions.NotFoundException;
 import fr.crafttogether.models.Bloc;
 import fr.crafttogether.repositories.BlocRepository;
 import lombok.AllArgsConstructor;
@@ -18,18 +19,24 @@ public class BlocService {
     }
 
     public Bloc findById(int id) {
-        return blocRepository.findById(id).orElse(null);
+        return blocRepository.findById(id).orElseThrow(() -> new NotFoundException("no bloc with id " + id + " exists"));
     }
+
+    public Bloc findByName(String name){ return blocRepository.findByName(name); }
 
     public Bloc save(Bloc bloc) {
         return blocRepository.save(bloc);
     }
 
     public void deleteById(int id) {
+        if(blocRepository.findById(id).isEmpty())
+            throw new NotFoundException("no bloc with id " + id + " exists");
         blocRepository.deleteById(id);
     }
 
     public Bloc update(Bloc bloc) {
+        if(blocRepository.findById(bloc.getId()).isEmpty())
+            throw new NotFoundException("no bloc with id " + bloc.getId() + " exists");
         return blocRepository.save(bloc);
     }
