+++ Craft2Gether/src/main/java/fr/crafttogether/services/ListeService.java
@@ -1,8 +1,10 @@
 package fr.crafttogether.services;
 
+import fr.crafttogether.exceptions.NotFoundException;
 import fr.crafttogether.models.Liste;
 import fr.crafttogether.repositories.ListeRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,21 +17,27 @@ public class ListeService {
     public List<Liste> findAll() {
         return listeRepository.findAll();
     }
-
     public Liste findById(int id) {
-        return listeRepository.findById(id).orElse(null);
+        return listeRepository.findById(id).orElseThrow(() -> new NotFoundException("no list with id " + id + " exists"));
     }
+    public Liste findByName(String nom){ return listeRepository.findByNom(nom); }
+
 
     public Liste save(Liste liste) {
         return listeRepository.save(liste);
     }
 
     public void deleteById(int id) {
+        if(listeRepository.findById(id).isEmpty())
+            throw new NotFoundException("no list with id " + id + " exists");
         listeRepository.deleteById(id);
     }
 
     public Liste update(Liste liste) {
+        if(listeRepository.findById(liste.getId()).isEmpty())
+            throw new NotFoundException("no list with id " + liste.getId() + " exists");
         return listeRepository.save(liste);
     }
+
 
 }
